requirements.txt
# langchain相关依赖
langchain>=0.1.0
langchain-openai~=0.1.25
pydantic~=2.10.6
langchain-core~=0.2.43

# 环境变量管理（用于保存API密钥）
python-dotenv==1.0.1

# FastAPI相关依赖
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
starlette~=0.44.0

# scripting和自动化相关依赖
pyautogui==0.9.54
Pillow==10.0.0
opencv-python==4.8.1.78
numpy~=1.26.4

# 网络请求相关依赖
requests~=2.32.5

# mysql相关依赖
PyMySQL~=1.1.2

# orm 框架依赖
SQLAlchemy~=2.0.43